# api/openapi/v1.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "labsnmp_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST adapter for the shared LabSNMP capability registry. Generated from internal/capabilities. Do not edit by hand.",
    "title": "LabSNMP Management API",
    "version": "v1",
    "x-generated-by": "internal/control/rest.RenderOpenAPI; DO NOT EDIT."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/state:validate": {
      "post": {
        "description": "Validate a candidate document and/or operations without writing.",
        "operationId": "state.validate",
        "parameters": [
          {
            "description": "Required when the labsnmp_session cookie authenticates a mutating request.",
            "in": "header",
            "name": "X-LabSNMP-CSRF",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Validate",
        "tags": [
          "state.validate"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "description": "Management listener (default address :8088)",
      "url": "/"
    }
  ]
}